1000원 미만 입력시 예외 처리

+++ src/main/java/seasonSix/lotto/common/utils/LottoGenerateUtil.java
@@ -6,12 +6,12 @@ import seasonSix.lotto.model.lotto.Lotto;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 public class LottoGenerateUtil {
-    public static List<Lotto> generate(int purchaseCount) {
+    public static List<Lotto> generate(long purchaseCount) {
         List<Lotto> lottos = new ArrayList<>();
-        IntStream.range(0, purchaseCount).forEach((i) -> {
+        LongStream.range(0, purchaseCount).forEach((i) -> {
             List<Integer> numbers = Randoms.pickUniqueNumbersInRange(1, 45, 6);
             numbers.sort(Comparator.naturalOrder());
             lottos.add(Lotto.newOne(numbers));
